update README, dependencies and typos; tag v3.2.0
