+++ src/deepquantum/__init__.py
@@ -3,7 +3,7 @@ This is the top level module from which all basic functions and classes of
 DeepQuantum can be directly imported.
 """
 
-__version__ = '3.1.1'
+__version__ = '3.2.0'
 
 
 from . import ansatz
