+++ src/deepquantum/photonic/__init__.py
@@ -7,12 +7,12 @@ from . import circuit
 from . import decompose
 from . import draw
 from . import gate
-from . import hafnian
+from . import hafnian_
 from . import mapper
 from . import operation
 from . import qmath
 from . import state
-from . import torontonian
+from . import torontonian_
 
 from .ansatz import Clements, GaussianBosonSampling, GBS_Graph
 from .circuit import QumodeCircuit
@@ -20,11 +20,11 @@ from .decompose import UnitaryDecomposer
 from .draw import DrawClements
 from .gate import PhaseShift, BeamSplitter, MZI, BeamSplitterTheta, BeamSplitterPhi, BeamSplitterSingle, UAnyGate
 from .gate import Squeezing, Squeezing2, Displacement, DisplacementPosition, DisplacementMomentum
-from .hafnian import hafnian
+from .hafnian_ import hafnian
 from .mapper import UnitaryMapper
 from .qmath import permanent, takagi, xxpp_to_xpxp, xpxp_to_xxpp, quadrature_to_ladder, ladder_to_quadrature
 from .state import FockState, GaussianState
-from .torontonian import torontonian
+from .torontonian_ import torontonian
 from .utils import set_hbar, set_kappa
 
 hbar = 2
