+++ src/deepquantum/photonic/hafnian_.py
